refactor(api): narrow the public API to the consumer-facing surface
2.0.0 is the moment to lock the public API, so restrict it to what
downstream consumers reasonably need and stop exporting low-level
implementation details. Verified against the OMEMO 2 integration in
converse.js, which uses only SessionBuilder/SessionCipher/KeyHelper/
OMEMOAddress and curvePubKeyToEd25519PubKey at runtime (WebCrypto handles
its HKDF/AES/hashing).

Removed from the public API (still available internally):
- raw crypto primitives: encrypt, decrypt, sign, hash, HKDF, HKDFInternal,
  verifyMAC, ECDHE, getRandomBytes, createKeyPair, Ed25519Sign, Ed25519Verify
- internalCrypto and the Curve25519 class
- getProtocolProfile and the ProtocolProfile type (internal protocol machinery)

Kept: SessionBuilder, SessionCipher, KeyHelper, OMEMOAddress,
FingerprintGenerator, SessionRecord, InMemoryStore, startWorker/stopWorker,
util, the Ed25519<->Curve conversion helpers (needed for omemo:2 fingerprints),
and the contract types/enums.

BREAKING CHANGE: the low-level crypto primitives, internalCrypto, the
Curve25519 class, getProtocolProfile and ProtocolProfile are no longer
exported from the package entry point.
